@@ -2,18 +2,21 @@ package com.deviousindustries.testtask;
 
 import android.content.Intent;
 import android.database.Cursor;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.TextView;
 
+import com.deviousindustries.testtask.Classes.Task;
+import com.deviousindustries.testtask.Classes.Time;
+
 public class Details_Session extends AppCompatActivity{
 
     ArrayListContainer mSessionTaskList;
     TimeKeeper timeKeeper;
-    Intent mIntent;
+    Intent fintent;
     Time mTime;
 
     @Override
@@ -25,8 +28,8 @@ public class Details_Session extends AppCompatActivity{
         timeKeeper.setMode(2);
         mTime = new Time();
 
-        mIntent = getIntent();
-        Bundle extras = mIntent.getExtras();
+        fintent = getIntent();
+        Bundle extras = fintent.getExtras();
         if (extras != null){
             mTime = new Time(getIntent().getLongExtra("EXTRA_TIME_ID",-1));
         }
@@ -49,21 +52,21 @@ public class Details_Session extends AppCompatActivity{
     @Override
     protected void onResume(){
         super.onResume();
-        if (mTime.mlngTimeID != -1){
+        if (mTime.flngTimeID != -1){
             LoadSession();
         }
     }
 
     private void LoadSession() {
-        setSessionTitle(mTime.mstrTitle);
+        setSessionTitle(mTime.fstrTitle);
         timeKeeper.loadTimeKeeper(mTime);
 
-        try(Cursor tblTask = mTime.getTasks()){
+        try(Cursor tblTask = DatabaseAccess.findOneOffs(mTime.flngTimeID)){
             mSessionTaskList.Clear();
             while (tblTask.moveToNext()){
                 Task tempTask = new Task(tblTask.getLong(tblTask.getColumnIndex("flngTaskID")));
-                mSessionTaskList.Add(tempTask.mstrTitle,
-                        tempTask.mlngTaskID);
+                mSessionTaskList.Add(tempTask.fstrTitle,
+                        tempTask.flngTaskID);
             }
         }
 
@@ -73,32 +76,32 @@ public class Details_Session extends AppCompatActivity{
     public void createSession (View view) {
         DatabaseAccess.mDatabase.beginTransaction();
         try {
-            if(mTime.mlngTimeID != -1){
+            if(mTime.flngTimeID != -1){
                 mTime.clearGenerationPoints();
             }
-            mTime = timeKeeper.createTimeDetails(mTime.mlngTimeID,
-                    mTime.mintTimeframe,
-                    mTime.mlngTimeframeID,
+            mTime = timeKeeper.createTimeDetails(mTime.flngTimeID,
+                    mTime.fintTimeframe,
+                    mTime.flngTimeframeID,
                     true,
                     getSessionTitle());
             mTime.refreshInstances();
 
             //Because session changed (possibly) we need to update any oneoffs that haven't yet been completed.
-            try(Cursor oneOffs = mTime.findOneOffs()) {
+            try(Cursor oneOffs = DatabaseAccess.findOneOffs(mTime.flngTimeID)) {
                 while (oneOffs.moveToNext()) {
                     Task tempTask = new Task(oneOffs.getLong(oneOffs.getColumnIndex("flngTaskID")));
-                    new Time(tempTask.mlngTimeID).clearGenerationPoints();
-                    Time tempTime = mTime.createOneOff(tempTask.mlngTimeID);//because we're supplying the time ID we shouldn't need to replace the session id
+                    new Time(tempTask.flngTimeID).clearGenerationPoints();
+                    Time tempTime = mTime.createOneOff(tempTask.flngTimeID);//because we're supplying the time ID we shouldn't need to replace the session id
 
                     tempTask.finishActiveInstances(3);
-                    tempTime.generateInstances(true, tempTask.mlngTaskID);
+                    tempTime.generateInstances(true, tempTask.flngTaskID);
                 }
             }
 
             DatabaseAccess.mDatabase.setTransactionSuccessful();
 
-            mIntent.putExtra("EXTRA_SESSION_ID", mTime.mlngTimeID);
-            setResult(RESULT_OK, mIntent);
+            fintent.putExtra("EXTRA_SESSION_ID", mTime.flngTimeID);
+            setResult(RESULT_OK, fintent);
             finish();
         } catch (Exception e) {
             e.printStackTrace();
